% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{aic}
\alias{aic}
\title{Compute the \eqn{AIC} based on \eqn{SSE}}
\usage{
aic(fitobj)
}
\arguments{
\item{fitobj}{Output of the \code{\link[discounting]{fit}} function.}
}
\value{
Numeric denoting the \eqn{AIC}
}
\description{
Compute the \eqn{AIC} based on \eqn{SSE}
}
\examples{
# Generate a dataset
data <- dataset(Y = rnorm(100), X = rnorm(100)) |>
  suppressWarnings()

# Estimate the exponential discounting model on these data
fitobj <- fit(
  exponential(),
  data,
  itermax = 50,
  trace = FALSE
)

# Compute the AIC based on this fitobj
aic(fitobj)

}
