% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{autocorrelation}
\alias{autocorrelation}
\title{Compute the autocorrelation of the residuals}
\usage{
autocorrelation(fitobj, lag = 1)
}
\arguments{
\item{fitobj}{Output of the \code{\link[discounting]{fit}} function.}

\item{lag}{The lag of the autocorrelation to compute. Defaults to \code{1}.}
}
\value{
Numeric denoting the autocorrelation of the residuals. If multiple
dimensions exist in the residuals, then the autocorrelation will be computed
for each dimension separately and a mean of them will be returned.
}
\description{
Compute the autocorrelation of the residuals
}
\examples{
# Generate a dataset
data <- dataset(Y = rnorm(100), X = rnorm(100)) |>
  suppressWarnings()

# Estimate the exponential discounting model on these data
fitobj <- fit(
  exponential(),
  data,
  itermax = 50,
  trace = FALSE
)

# Compute the autocorrelation based on this fitobj
autocorrelation(fitobj)

}
