% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bias}
\alias{bias}
\title{Compute the mean of the residuals}
\usage{
bias(fitobj)
}
\arguments{
\item{fitobj}{Output of the \code{\link[discounting]{fit}} function.}
}
\value{
Numeric denoting the mean of the residuals. If multiple
dimensions exist in the residuals, then the mean will be computed
for each dimension separately and the average of them will be returned.
}
\description{
Compute the mean of the residuals
}
\examples{
# Generate a dataset
data <- dataset(Y = rnorm(100), X = rnorm(100)) |>
  suppressWarnings()

# Estimate the exponential discounting model on these data
fitobj <- fit(
  exponential(),
  data,
  itermax = 50,
  trace = FALSE
)

# Compute the autocorrelation based on this fitobj
bias(fitobj)

}
