% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bic}
\alias{bic}
\title{Compute the \eqn{BIC} based on \eqn{SSE}}
\usage{
bic(fitobj)
}
\arguments{
\item{fitobj}{Output of the \code{\link[discounting]{fit}} function.}
}
\value{
Numeric denoting the \eqn{BIC}
}
\description{
Compute the \eqn{BIC} based on \eqn{SSE}
}
\examples{
# Generate a dataset
data <- dataset(Y = rnorm(100), X = rnorm(100)) |>
  suppressWarnings()

# Estimate the exponential discounting model on these data
fitobj <- fit(
  exponential(),
  data,
  itermax = 50,
  trace = FALSE
)

# Compute the BIC based on this fitobj
bic(fitobj)

}
