% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_bound_length}
\alias{check_bound_length}
\title{Check performed within the \code{\link[discounting]{get_bounds}} function}
\usage{
check_bound_length(n, lower, upper)
}
\arguments{
\item{n}{Integer denoting the needed length of the bounds}

\item{lower, upper}{Numeric vector containing the user-specified lower and 
upper bounds for which the length should be checked.}
}
\value{
List containing the corrected lower and upper bounds (in that order)
}
\description{
Check performed within the \code{\link[discounting]{get_bounds}} function
}
\keyword{internal}
