% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_covariance}
\alias{count_covariance}
\title{Count the number of parameters in the covariance matrix}
\usage{
count_covariance(d, covariance = "symmetric")
}
\arguments{
\item{d}{Integer denoting the dimensionality of the model.}

\item{covariance}{Character denoting the structure of the covariance matrix.
Can either be \code{"symmetric"} (symmetric around the diagonal) or 
\code{"isotropic"} (diagonal). Defaults to \code{"symmetric"}.}
}
\value{
Integer denoting the number of parameters in the covariance matrix
under the specified conditions.
}
\description{
Count the number of parameters in the covariance matrix
}
\examples{
count_covariance(2, "symmetric")
count_covariance(2, "isotropic")

}
