% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_parameters}
\alias{count_parameters}
\alias{count_parameters,exponential-method}
\title{Count number of parameters for a model}
\usage{
count_parameters(model, ...)

\S4method{count_parameters}{exponential}(
  model,
  dynamics = "isotropic",
  covariance = "symmetric",
  parameters_only = FALSE
)
}
\arguments{
\item{model}{An instance of the \code{\link[discounting]{model-class}}}

\item{...}{Arguments passed on to the methods.}

\item{dynamics}{Character denoting the structure of the dynamical matrices.
Can either be \code{"anisotropic"} (completely free), \code{"symmetric"}
(symmetric around the diagonal), and \code{"isotropic"} (diagonal). Note that
this influences different parameters for different models, namely 
\eqn{\Gamma} for the exponential discounting model, \eqn{N} and \eqn{K} for
the quasi-hyperbolic discounting model, and \eqn{\Gamma} and \eqn{N} for the
double-exponential discounting model. Defaults to \code{"isotropic"}.}

\item{covariance}{Character denoting the structure of the covariance matrix.
Can either be \code{"symmetric"} (symmetric around the diagonal) or 
\code{"isotropic"} (diagonal). Defaults to \code{"isotropic"}.}

\item{parameters_only}{Logical denoting whether to only count the number of 
parameters in de \code{parameter} slot of the model (\code{TRUE}), or to 
count the number of parameters in the covariance matrix as well 
(\code{FALSE}). Defaults to \code{FALSE}.}
}
\value{
Integer denoting the number of parameters the model contains within
the current specifications.
}
\description{
Count number of parameters for a model
}
\examples{
# Define a model with a particular dimensionality
my_model <- exponential(
  parameters = list(
    "alpha" = numeric(2),
    "beta" = matrix(0, nrow = 2, ncol = 5),
    "gamma" = matrix(0, nrow = 2, ncol = 2)
  ),
  covariance = matrix(0, nrow = 2, ncol = 2)
)

# Get the number of parameters for this model under no restrictions (i.e., 
# anisotropic forgetting factors and symmetric covariances)
count_parameters(
  my_model, 
  dynamics = "anisotropic",
  covariance = "symmetric"
)

# Get the number of parameters for this model in the most limited case 
# (i.e., isotropic forgetting factors and covariances)
count_parameters(
  my_model,
  dynamics = "isotropic",
  covariance = "isotropic"
)

}
