% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{class}
\name{dataset-class}
\alias{dataset-class}
\title{Dataset Class}
\description{
This class contains the standard format for the dataset to be used within this 
project, ensuring a standardized approach for all computations.
}
\section{Slots}{

\describe{
\item{\code{Y}}{Numeric matrix of observations to be fitted, that is the dependent 
variables}

\item{\code{X}}{Numeric matrix of the stimuli that should be related to \code{Y}, 
that is the independent variables}
}}

\seealso{
\code{\link[discounting]{dataset}}
}
