% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset}
\alias{dataset}
\title{Constructor for the \code{\link[discounting]{dataset-class}}}
\usage{
dataset(
  data = NULL,
  y_cols = NULL,
  x_cols = NULL,
  Y = NULL,
  X = NULL,
  sorting_variable = NULL
)
}
\arguments{
\item{data}{A \code{data.frame} containing the dependent and independent 
variables of interest for a single person. Defaults to \code{NULL}, in which
case an empty instance of the \code{\link[discounting]{dataset-class}} will 
be created.}

\item{y_cols}{A character vector denoting the names of the columns for the 
dependent variables. Defaults to \code{NULL}, in which case an error will 
be thrown asking for more information.}

\item{x_cols}{A character vector denoting the names of the columns for the 
independent variables. Defaults to \code{NULL}, in which case an error will 
be thrown asking for more information.}

\item{Y}{Numeric vector or numeric matrix containing the values for the 
dependent variable(s). Defaults to \code{NULL}, in which case it will 
become an empty matrix of length \eqn{N x 1}. Ignored if \code{data} is 
specified instead.}

\item{X}{Numeric vector or numeric matrix containing the values for the 
independent variable(s). Defaults to \code{NULL}, in which case the 
\code{data} should be specified instead.}

\item{sorting_variable}{Character denoting the columnn name of a variable 
in the \code{data} according to which the values for \code{Y} and \code{X}
should be sorted. Defaults to \code{NULL}, meaning that no sorting should 
happen.}
}
\value{
Instance of the \code{\link[discounting]{dataset-class}}.
}
\description{
Takes in a \code{data.frame} and, based on the value of the arguments, 
constructs an instance of the \code{\link[discounting]{dataset-class}}. Through
this type of class definition, it is ensured that the estimation, simulation,
cross-validation, and recovery routines can all make use of the same 
data structure, making their definition easier within this package. Note 
that this class is only defined on the within-subject level, meaning that 
the data provided to this constructor should be on this level.
}
\examples{
# Create a data.frame for reference
my_data <- data.frame(
  DV_1 = rep(1, each = 10),
  DV_2 = rep(2, each = 10),
  IV_1 = rep(3, each = 10),
  IV_2 = rep(4, each = 10),
  IV_3 = rep(5, each = 10)
)

# Create a dataset with only a single DV and IV
dataset(
  data = my_data, 
  y_cols = "DV_1",
  x_cols = "IV_1"
)

# Create a dataset with multiple DV and IV
dataset(
  data = my_data,
  y_cols = c("DV_1", "DV_2"),
  x_cols = c("IV_1", "IV_2", "IV_3")
)
 
}
\seealso{
\code{\link[discounting]{dataset-class}}
}
