% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discounting-package.R
\docType{package}
\name{discounting-package}
\alias{discounting}
\alias{discounting-package}
\title{discounting: Discounting of Affect}
\description{
Functions created to aid with a large-scale evaluation of discounting 
models for capturing affect in experimental and daily-life datasets. 
Introduces a class for each separate model as well as a class for datasets,
making every function internal to the package run within a standardized 
framework
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ndpvh/discounting-affect}
  \item Report bugs at \url{https://github.com/ndpvh/discounting-affect/issues}
}

}
\keyword{internal}
