% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{class}
\name{double_exponential-class}
\alias{double_exponential-class}
\title{Double-exponential Class}
\description{
Class for the double-exponential discounting model. Follows a similar 
structure to the \code{\link[discounting]{model-class}}.
}
\section{Slots}{

\describe{
\item{\code{d}}{Integer denoting the number of dimensions of the model}

\item{\code{k}}{Integer denoting the number of independent variables}

\item{\code{parameters}}{List containing the parameters relevant to a particular 
model instance}

\item{\code{covariance}}{Numeric matrix denoting the residual covariance of the 
model}
}}

\seealso{
\code{\link[discounting]{double_exponential}}
}
