% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{double_exponential}
\alias{double_exponential}
\title{Constructor for the \code{\link[discounting]{double_exponential-class}}}
\usage{
double_exponential(
  d = NA,
  k = NA,
  parameters = NULL,
  covariance = NULL,
  cholesky = FALSE
)
}
\arguments{
\item{d}{Integer denoting the number of dimensions of the model. Defaults to 
\code{NA}, in which case this dimensionality is inferred from the parameters.
Note that if \code{d} and \code{k} are both defined while \code{parameters}
is left unspecified, then the model will automatically create an empty 
model of the specified dimensionality.}

\item{k}{Integer denoting the number of independent variables. Defaults to 
\code{NA}, in which case this dimensionality is inferred from the parameters.}

\item{parameters}{List containing the parameters relevant to a particular 
model instance. Defaults to \code{NULL}, creating an empty model with 
\eqn{d = 1} and \eqn{k = 1} (unless \code{d} and \code{k} are specified).}

\item{covariance}{Numeric matrix denoting the residual covariance of the 
model. Default to \code{NULL}, creating a matrix of \code{0}s with the 
dimensionality implied by \code{d}.}

\item{cholesky}{Logical denoting whether \code{covariance} is a 
lower-triangular decomposition matrix instead of an actual covariance matrix.
Defaults to \code{FALSE}.}
}
\value{
Instance of \code{\link[discounting]{double_exponential-class}}
}
\description{
Defines an instance of the \code{\link[discounting]{double_exponential-class}}, 
that is the class defining the double-exponential discounting model. For the 
mathematical equations and how to specify them in this constructor, users can 
look at the details. For more information on the model itself, see the 
vignette on the model definitions.
}
\details{
The double-exponential discounting model assumes that the effect of stimuli on 
affect fades away at an exponential rate as constructed through two processes,
one representing longer and the other slower discounting rates, where each 
process' effect is accounted for through a weighted sum. The model can be 
defined as:

\deqn{\boldsymbol{y}_t = \boldsymbol{\alpha} + \sum_{j = 0}^t \left( 
\omega \Gamma^j + (1 - \omega) N^j \right) B \boldsymbol{x}_{t - j} + 
\boldsymbol{\epsilon}_t}

where \eqn{\boldsymbol{\alpha}} is a \eqn{d}-dimensional vector 
representing the mean of the process \eqn{\boldsymbol{y}}, \eqn{\Gamma} and 
\eqn{N} are \eqn{d \times d} matrices containing the forgetting factors, 
determining how long the effect of the independent variables 
\eqn{\boldsymbol{x}} on the process \eqn{\boldsymbol{y}} lingers on and 
therefore determining the dynamics of the system, \eqn{\omega \in [0, 0.5]}
determines the strength with which \eqn{\Gamma} and \eqn{N} influence the 
discounting process, \eqn{B} is a \eqn{d \times k}matrix containing the 
slopes of the independent variables \eqn{\boldsymbol{x}}, and 
\eqn{\boldsymbol{\epsilon}} represents the residuals of the system. 
Within this package, we assume that:

\deqn{\boldsymbol{\epsilon} \overset{iid}{\sim} N(\boldsymbol{0}, \Sigma)}

where \eqn{\Sigma} is a \eqn{d \times d} matrix representing the residual 
covariance matrix. 

To define the model, one should minimally define the parameters of the model
through \code{parameters} and \code{covariance}. In \code{covariance}, one 
either provides the \eqn{d \times d} residual covariance matrix \eqn{\Sigma}
(\code{cholesky = FALSE}) or a lower-triangular \eqn{d \times d} matrix 
\eqn{G} that represents part of the decomposition of this matrix 
(\code{cholesky = TRUE}), so that we can retrieve the covariance matrix as:

\deqn{\Sigma = G G^T}

The latter option ensures that the covariance matrix \eqn{\Sigma} defined 
within the model is positive-definite, but may require some additional 
thinking on the side of the user.

In \code{parameters}, one should provide a named list with instances 
\code{"alpha"}, \code{"beta"}, \code{"omega"}, \code{"gamma"}, and 
\code{"nu"}, each defining the respective parameters 
\eqn{\boldsymbol{\alpha}}, \eqn{B}, \eqn{\omega}, \eqn{\Gamma} and \eqn{N} 
of the model. Note that the dimensionalities of these parameters should match 
up and are actively checked when constructing this class. 

There are some important restrictions to take into account when defining the 
model:
\itemize{
  \item{The eigenvalues \eqn{\lambda_i} of \eqn{\Gamma} and \eqn{N} should
        all lie between 0 and 1, or \eqn{\lambda_i \in [0, 1)}, a restriction
        that imposes exponential decay rather than sawtooth or more complex
        dynamics}
  \item{The weight \eqn{\omega} should lie between 0 and 0.5, or 
        \eqn{\omega \in [0, 0.5]}, a restriction that is imposed to ensure 
        recoverability of the model}
  \item{The covariance matrix \eqn{\Sigma} should be positive definite, as 
        explained above}
}
}
\examples{
double_exponential(
  d = 2,
  k = 2,
  parameters = list(
    "alpha" = numeric(2),
    "beta" = diag(2) * 2,
    "omega" = 0.25,
    "gamma" = diag(2) * 0.75,
    "nu" = diag(2) * 0.5
  ),
  covariance = diag(2)
)

double_exponential(
  d = 2,
  k = 2
)

}
\seealso{
\code{\link[discounting]{model-class}}
\code{\link[discounting]{exponential-class}}
\code{\link[discounting]{quasi_hyperbolic-class}}
\code{\link[discounting]{double_exponential-class}}
}
