% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{class}
\name{exponential-class}
\alias{exponential-class}
\title{Exponential Class}
\description{
Class for the exponential discounting model. Follows a similar structure to
the \code{\link[discounting]{model-class}}.
}
\section{Slots}{

\describe{
\item{\code{d}}{Integer denoting the number of dimensions of the model}

\item{\code{k}}{Integer denoting the number of independent variables}

\item{\code{parameters}}{List containing the parameters relevant to a particular 
model instance}

\item{\code{covariance}}{Numeric matrix denoting the residual covariance of the 
model}
}}

\seealso{
\code{\link[discounting]{exponential}}
}
