% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\alias{fit,model,dataset-method}
\title{Fit a model to data}
\usage{
fit(model, data, ...)

\S4method{fit}{model,dataset}(
  model,
  data,
  dynamics = "isotropic",
  covariance = "symmetric",
  optimizer = "optim",
  lower = NULL,
  upper = NULL,
  method = "Nelder-Mead",
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{model}{Instance of the \code{\link[discounting]{model-class}}, 
defining the model to evaluate the objective function for.}

\item{data}{Instance of the \code{\link[discounting]{dataset-class}}
containing the data to fit the model to.}

\item{...}{Arguments passed on to the control parameters of the optimizer, 
either to \code{\link[DEoptim]{DEoptim.control}}, the \code{opts} 
argument of \code{\link[nloptr]{nloptr}}, the \code{control} list of the 
\code{\link[stats]{optim}} function, or to the additional arguments of your
own \code{optimizer} function (see Details).}

\item{dynamics}{Character denoting the structure of the dynamical matrices.
Can either be \code{"anisotropic"} (completely free), \code{"symmetric"}
(symmetric around the diagonal), and \code{"isotropic"} (diagonal). Note that
this influences different parameters for different models, namely 
\eqn{\Gamma} for the exponential discounting model, \eqn{N} and \eqn{K} for
the quasi-hyperbolic discounting model, and \eqn{\Gamma} and \eqn{N} for the
double-exponential discounting model. Defaults to \code{"isotropic"}.}

\item{covariance}{Character denoting the structure of covariance matrix.
Can either by \code{"symmetric"} (symmetric around the diagonal) and 
\code{"isotropic"} (diagonal). Defaults to \code{"symmetric"}.}

\item{optimizer}{Character denoting the optimizer to use for the estimation.
Can either be \code{"DEoptim"} for the differential evolution algorithm in 
\code{\link[DEoptim]{DEoptim}}, \code{"nloptr"} for the library implemented
in \code{\link[nloptr]{nloptr}}, or \code{"optim"} for the default 
optimization library in R. Defaults to \code{"optim"}. Alternatively, 
\code{optimizer} can also be a custom function that defines the estimation 
routine. To see what this function should look like, we refer the interested
reader to the details.}

\item{lower, upper}{Numeric vector containing lower and upper bounds for the 
parameters in the estimation routine. Uses the same defaults as 
\code{\link[discounting]{get_bounds}}.}

\item{method}{Character denoting the method to be used for estimation. Ignored
if \code{optimizer} is not \code{"optim"}, in which case the native arguments
of the other optimization packages can be used.}

\item{na.rm}{Logical denoting whether to remove NAs in the calculation of the 
sum squared error. Defaults to \code{TRUE}.}
}
\value{
An named list containing an instance of the 
\code{\link[discounting]{model-class}} with the estimated parameters 
(\code{"model"}), the results of the optimization procedure (\code{"fit"}), 
the value of the objective after ending the optimization procedure 
(\code{"objective"}), the residuals of the model (\code{"residuals"}), and a 
named vector containing the values of the estimated parameters linked to a 
character vector explaining their content (\code{"parameters"}).
}
\description{
This method enables the user to fit a model (defined in \code{model}) on a
particular dataset (defined in \code{data}). The estimation procedure makes
use of numerical optimization using either \code{\link[DEoptim]{DEoptim}} or
\code{\link[nloptr]{nloptr}}, as specified by the user. Estimation proceeds
through an optimization according to the output of the objective function 
of the provided model, as defined through 
\code{\link[discounting]{objective_function}}, thus using least-squares
as optimization standard.
}
\details{
Note that currently, least-squares estimation is assumed, meaning that the 
covariance matrix is left out of the objective function. If maximum-likelihood
were needed instead, then this function would need to change.

By default, the \code{fit} function performs optimizations through the 
\code{optim} function in R, alternatively allowing users to make use of the
\code{DEoptim} or \code{nloptr} packages instead. For more difficult models,
however, it may be useful to have greater control over the optimization 
procedure, for example by specifying your own initial conditions, by using 
a combination of different optimization algorithms at once, or by using 
another library not provided through this package. 

In these cases, one can provide a function to the argument \code{optimizer}
instead of a string. This function should at least take in 3 argument: A
function representing the objective function, a lower bound for the 
parameters, and an upper bound for these same parameters (cf. \code{optim}, 
\code{DEoptim}, and \code{nloptr}). This function can furthermore take in 
additional argument, for example specifying the maximal number of iterations
or controlling other aspects of the estimation routine. These can be specified
in the \code{fit} function through the optional arguments \code{...}.

The custom optimization function should, minimally, output a named list with 
slots \code{"parameters"} and \code{"objective"}, containing the final 
result of the parameter estimation and the value of the objective function 
for this final parameter set respectively. These will then be used by the 
\code{fit} function as final results of the optimization procedure.
}
\examples{
# Simulate data to use for this example
data <- simulate(
  quasi_hyperbolic(
    parameters = list(
      "alpha" = c(1, -1) ,
      "beta" = matrix(2, nrow = 2, ncol = 2),
      "nu" = diag(2) * 0.75,
      "kappa" = diag(2) * 0.5
    ),
    covariance = matrix(c(1, 0.25, 0.25, 1), nrow = 2, ncol = 2)
  ),
  X = matrix(rnorm(200), nrow = 100, ncol = 2)
)

# Evaluate the objective function for an exponential model with a particular
# set of parameters
fit(
  exponential(d = 2, k = 2),
  data,
  dynamics = "isotropic",
  covariance = "isotropic",
  optimizer = "DEoptim",
  itermax = 50,
  trace = FALSE
)

}
\concept{estimation}
