% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_parameters}
\alias{generate_parameters}
\alias{generate_parameters,model-method}
\title{Generate model parameters}
\usage{
generate_parameters(model, ...)

\S4method{generate_parameters}{model}(model, ...)
}
\arguments{
\item{model}{Instance of the \code{\link[discounting]{model-class}}}

\item{...}{Additional arguments passed on to 
\code{\link[discounting]{get_bounds}}}
}
\value{
Numeric vector of parameters for the model.
}
\description{
Generation of model parameters within this function always occurs according 
to a uniform distribution within the bounds defined for the models.
}
\examples{
generate_parameters(
  double_exponential(d = 2, k = 3),
  dynamics = "isotropic",
  covariance = "isotropic",
  parameters_only = FALSE
)

}
