% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_bounds}
\alias{get_bounds}
\alias{get_bounds,exponential-method}
\alias{get_bounds,quasi_hyperbolic-method}
\alias{get_bounds,double_exponential-method}
\title{Get bounds of the parameters of a model}
\usage{
get_bounds(model, ...)

\S4method{get_bounds}{exponential}(
  model,
  dynamics = "isotropic",
  covariance = "symmetric",
  parameters_only = TRUE,
  lower = NULL,
  upper = NULL
)

\S4method{get_bounds}{quasi_hyperbolic}(
  model,
  dynamics = "isotropic",
  covariance = "symmetric",
  parameters_only = TRUE,
  lower = NULL,
  upper = NULL
)

\S4method{get_bounds}{double_exponential}(
  model,
  dynamics = "isotropic",
  covariance = "symmetric",
  parameters_only = TRUE,
  lower = NULL,
  upper = NULL
)
}
\arguments{
\item{model}{Instance of the \code{\link[discounting]{model-class}}.}

\item{...}{Additional arguments passed on to the methods.}

\item{dynamics}{Character denoting the structure of the dynamical matrices.
Can either be \code{"anisotropic"} (completely free), \code{"symmetric"}
(symmetric around the diagonal), and \code{"isotropic"} (diagonal). Note that
this influences different parameters for different models, namely 
\eqn{\Gamma} for the exponential discounting model, \eqn{N} and \eqn{K} for
the quasi-hyperbolic discounting model, and \eqn{\Gamma} and \eqn{N} for the
double-exponential discounting model. Defaults to \code{"isotropic"}.}

\item{covariance}{Character denoting the structure of the covariance matrix.
Can either be \code{"symmetric"} (symmetric around the diagonal) or 
\code{"isotropic"} (diagonal). Defaults to \code{"symmetric"}.}

\item{parameters_only}{Logical denoting whether to only get bounds for the 
parameters in de \code{parameter} slot of the model (\code{TRUE}), or to 
fill the covariance matrix as well (\code{FALSE}). Defaults to \code{TRUE}.}

\item{lower}{Numeric vector of the same length as the substantive parameters
in the model, denoting the lower bounds for each one. For the exponential
discounting model, for example, a value of `c(0, 0, 0, 0)` would denote a 
minimal value of 0 for all values in \eqn{\bm{\alpha}}, \eqn{B}, \eqn{\Gamma},
and \eqn{\Sigma}. Defaults depend on the model itself and have been chosen 
while accounting for natural bounds in the data used in this project. Note
that if \code{parameters_only = TRUE}, there is no need to specify a bound 
for the covariances.}

\item{upper}{Numeric vector that functions the same as \code{lower}, but 
specifying the upper bounds instead.}
}
\value{
Named list containing numeric vectors denoting the lower and upper 
bounds for the parameters (under \code{"lower"} and \code{"upper"} resp.).
The order maintained in this
}
\description{
Get bounds of the parameters of a model
}
\examples{
# Create an empty instance of the exponential discounting model
my_model <- exponential()

# Get the bounds for the model
get_bounds(
  my_model,
  dynamics = "anisotropic",
  covariance = "isotropic",
  parameters_only = FALSE
)

# Get the bounds for the model when specifying your personal lower and upper
# bounds
get_bounds(
  my_model,
  dynamics = "anisotropic",
  covariance = "isotropic",
  parameters_only = FALSE,
  lower = c(0, -100, 0, 0.01),
  upper = c(1, 100, 0.5, 1)
)

}
