% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_bounds_covariance}
\alias{get_bounds_covariance}
\title{Get bounds for the parameters in the covariance matrix}
\usage{
get_bounds_covariance(d, lower, upper, covariance = "symmetric")
}
\arguments{
\item{d}{Integer denoting the dimensionality of the model.}

\item{lower, upper}{Numeric value containing the value of the lower and 
upper bounds for the covariance matrix.}

\item{covariance}{Character denoting the structure of the covariance matrix.
Can either be \code{"symmetric"} (symmetric around the diagonal) or 
\code{"isotropic"} (diagonal). Defaults to \code{"symmetric"}.}
}
\value{
List containing two numeric vectors of lower and upper bounds for the 
covariance matrix in that order.
}
\description{
Get bounds for the parameters in the covariance matrix
}
\examples{
count_covariance(
  2, 
  0,
  1,
  covariance = "symmetric"
)

count_covariance(
  2,
  0, 
  1,
  covariance = "isotropic"
)

}
