% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{index}
\alias{index}
\alias{index,exponential-method}
\alias{index,quasi_hyperbolic-method}
\alias{index,double_exponential-method}
\title{Index model parameters}
\usage{
index(model, ...)

\S4method{index}{exponential}(
  model,
  dynamics = "isotropic",
  parameters_only = TRUE,
  full = TRUE,
  cholesky = TRUE,
  ...
)

\S4method{index}{quasi_hyperbolic}(
  model,
  dynamics = "isotropic",
  parameters_only = TRUE,
  full = TRUE,
  cholesky = TRUE,
  ...
)

\S4method{index}{double_exponential}(
  model,
  dynamics = "isotropic",
  parameters_only = TRUE,
  full = TRUE,
  cholesky = TRUE,
  ...
)
}
\arguments{
\item{model}{Instance of the \code{\link[discounting]{model-class}}}

\item{...}{Additional arguments passed on to the methods.}

\item{dynamics}{Character denoting the structure of the dynamical matrices.
Can either be \code{"anisotropic"} (completely free), \code{"symmetric"}
(symmetric around the diagonal), and \code{"isotropic"} (diagonal). Note that
this influences different parameters for different models, namely 
\eqn{\Gamma} for the exponential discounting model, \eqn{N} and \eqn{K} for
the quasi-hyperbolic discounting model, and \eqn{\Gamma} and \eqn{N} for the
double-exponential discounting model. Defaults to \code{"isotropic"}.}

\item{parameters_only}{Logical denoting whether to only fill the 
parameters in de \code{parameter} slot of the model (\code{TRUE}), or to 
fill the covariance matrix as well (\code{FALSE}). Defaults to \code{TRUE}.}

\item{full}{Logical denoting whether to provide the full matrices or 
whether a partial fill is sufficient. Makes the distinction between a 
symmetric matrix being lower-triangular or fully filled. Defaults to 
\code{TRUE} as for most purposes, you would want the complete matrix. An 
example of where this is \code{FALSE} can be found in 
\code{\link[discounting]{parameters}}. Note that if \code{full = FALSE} and 
\code{parameters_only = FALSE}, this will automatically let the 
\code{cholesky} argument become \code{TRUE}, ensuring only the 
lower-triangular of the covariance matrix is indexed (as required by 
\code{full}).}

\item{cholesky}{Logical denoting whether the idea is to use the Cholesky 
decomposition to create the values of the covariance matrix. In this case, 
the indices should only span the lower-triangular of the matrix. Defaults 
to \code{TRUE}.}
}
\value{
Instance of the \code{\link[discounting]{model-class}} with the 
parameters being indexed
}
\description{
Fill the model parameters with indices denoting which values to use for a 
particular model.
}
\examples{
index(
  double_exponential(d = 2, k = 3),
  full = TRUE
)

index(
  double_exponential(d = 2, k = 3),
  full = FALSE
)

}
