% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{index_covariance}
\alias{index_covariance}
\title{Index the covariance matrix}
\usage{
index_covariance(d, start, covariance = "symmetric", cholesky = TRUE)
}
\arguments{
\item{d}{Integer denoting the dimensionality of the model.}

\item{start}{Integer denoting the index at which to start the indexing of 
the covariance matrix, so that it is the first value of the index.}

\item{covariance}{Character denoting the structure of the covariance matrix.
Can either be \code{"symmetric"} (symmetric around the diagonal) or 
\code{"isotropic"} (diagonal). Defaults to \code{"symmetric"}.}

\item{cholesky}{Logical denoting whether the idea is to use the Cholesky 
decomposition to create the values of the covariance matrix. In this case, 
the indices should only span the lower-triangular of the matrix. Defaults 
to \code{TRUE}.}
}
\value{
Indexed covariance matrix.
}
\description{
Index the covariance matrix
}
\examples{
index_covariance(
  2,
  6,
  covariance = "symmetric",
  cholesky = FALSE
)

index_covariance(
  2,
  6,
  covariance = "isotropic",
  cholesky = FALSE
)

}
