% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{class}
\name{model-class}
\alias{model-class}
\title{Model Class}
\description{
Class that created the general structure of all models introduced within this 
package.
}
\section{Slots}{

\describe{
\item{\code{d}}{Integer denoting the number of dimensions of the model}

\item{\code{k}}{Integer denoting the number of independent variables}

\item{\code{parameters}}{List containing the parameters relevant to a particular 
model instance}

\item{\code{covariance}}{Numeric matrix denoting the residual covariance of the 
model}
}}

