% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model}
\alias{model}
\title{Constructor for the \code{\link[discounting]{model-class}}}
\usage{
model(
  d = 1,
  k = 0,
  parameters = list(),
  covariance = matrix(0, nrow = d, ncol = d)
)
}
\arguments{
\item{d}{Integer denoting the number of dimensions of the model. Defaults to 
\code{1}.}

\item{k}{Integer denoting the number of independent variables. Defaults to 
\code{0}.}

\item{parameters}{List containing the parameters relevant to a particular 
model instance. Defaults to an empty list, meaning the model cannot be 
properly used.}

\item{covariance}{Numeric matrix denoting the residual covariance of the 
model. Default to a matrix of \code{0}s with the dimensionality implied by 
\code{d}}
}
\value{
Instance of \code{\link[discounting]{model-class}}
}
\description{
Constructor for the \code{\link[discounting]{model-class}}
}
\examples{
model(
  d = 2,
  k = 2,
  parameters = list(
    "delta" = numeric(2),
    "theta" = diag(2)
  ),
  covariance = diag(2)
)

}
\seealso{
\code{\link[discounting]{exponential-class}}
}
