% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{objective_function}
\alias{objective_function}
\alias{objective_function,model,dataset-method}
\title{Objective function}
\usage{
objective_function(model, data, parameters, ...)

\S4method{objective_function}{model,dataset}(model, data, parameters, dynamics = "isotropic")
}
\arguments{
\item{model}{Instance of the \code{\link[discounting]{model-class}}, 
defining the model to evaluate the objective function for.}

\item{data}{Instance of the \code{\link[discounting]{dataset-class}}
containing the data to fit the model to.}

\item{parameters}{Numeric vector containing the parameters of the model to 
compute the \eqn{SSE} for.}

\item{...}{Additional arguments passed on to the methods.}

\item{dynamics}{Character denoting the structure of the dynamical matrices.
Can either be \code{"anisotropic"} (completely free), \code{"symmetric"}
(symmetric around the diagonal), and \code{"isotropic"} (diagonal). Note that
this influences different parameters for different models, namely 
\eqn{\Gamma} for the exponential discounting model, \eqn{N} and \eqn{K} for
the quasi-hyperbolic discounting model, and \eqn{\Gamma} and \eqn{N} for the
double-exponential discounting model. Defaults to \code{"isotropic"}.}
}
\value{
The sum of squared error or \eqn{SSE}, quantifying misfit of the model
to the data.
}
\description{
This method defines how the objective function of the estimation procedure 
looks like. The result of this objective function is the sum squared error,
meaning the estimation routine makes use of a least-squares procedure.
}
\details{
For the purposes of the package, the objective function defines how the 
estimation routine should go about, therefore defining the procedure with 
which the fit of the model to the data under a particular set of parameters
is evaluated. Within this package, we use the least-squares procedure, 
therefore minimizing the sum of squared error (\eqn{SSE}), defined as:

\deqn{SSE = Y - \hat{Y}}

where \eqn{Y} represents the observed data and \eqn{\hat{Y}} the predictions
of the model under the proposed parameter set. 

The procedure as used here relies heavily on the implementation of the 
\code{\link[discounting]{predict}} method to get the predictions 
\eqn{\hat{Y}}. If a recovery does not go well, it may be useful to take a 
look at how this method is defined for your particular model.

Note that the objective function is only useful when using a non-analytic 
approach to estimation, as the one used in this package.
}
\examples{
# Simulate data to use for this example
data <- simulate(
  quasi_hyperbolic(
    parameters = list(
      "alpha" = 1, 
      "beta" = as.matrix(2),
      "nu" = as.matrix(0.75),
      "kappa" = as.matrix(0.5)
    ),
    covariance = as.matrix(1)
  ),
  X = rnorm(100)
)

# Evaluate the objective function for an exponential model with a particular
# set of parameters
objective_function(
  exponential(),
  data,
  c(1, 2, 0.75)
)

# Evaluate the objective function when using exactly the model and the 
# parameters that generated the data
objective_function(
  quasi_hyperbolic(),
  data,
  c(1, 2, 0.75, 0.5)
)

}
