% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parameter_names}
\alias{parameter_names}
\alias{parameter_names,exponential-method}
\alias{parameter_names,quasi_hyperbolic-method}
\alias{parameter_names,double_exponential-method}
\title{Name the relevant parameters of the model}
\usage{
parameter_names(model, ...)

\S4method{parameter_names}{exponential}(
  model,
  dynamics = "isotropic",
  covariance = "symmetric",
  parameters_only = TRUE
)

\S4method{parameter_names}{quasi_hyperbolic}(
  model,
  dynamics = "isotropic",
  covariance = "symmetric",
  parameters_only = TRUE
)

\S4method{parameter_names}{double_exponential}(
  model,
  dynamics = "isotropic",
  covariance = "symmetric",
  parameters_only = TRUE
)
}
\arguments{
\item{model}{Instance of the \code{\link[discounting]{model-class}}.}

\item{...}{Additional arguments passed on to the methods.}

\item{dynamics}{Character denoting the structure of the dynamical matrices.
Can either be \code{"anisotropic"} (completely free), \code{"symmetric"}
(symmetric around the diagonal), and \code{"isotropic"} (diagonal). Note that
this influences different parameters for different models, namely 
\eqn{\Gamma} for the exponential discounting model, \eqn{N} and \eqn{K} for
the quasi-hyperbolic discounting model, and \eqn{\Gamma} and \eqn{N} for the
double-exponential discounting model. Defaults to \code{"isotropic"}.}

\item{covariance}{Character denoting the structure of the covariance matrix.
Can either be \code{"symmetric"} (symmetric around the diagonal) or 
\code{"isotropic"} (diagonal). Defaults to \code{"symmetric"}.}

\item{parameters_only}{Logical denoting whether to only name the 
parameters in de \code{parameter} slot of the model (\code{TRUE}), or to 
fill the covariance matrix as well (\code{FALSE}). Defaults to \code{TRUE}.}
}
\value{
Character vector containing the names for the parameters in the 
model. The order corresponds to the assumptions of 
\code{\link[discounting]{get_bounds}} and \code{\link[discounting]{fill}}.
}
\description{
Name the relevant parameters of the model
}
\examples{
# Create an empty instance of the exponential discounting model
my_model <- exponential()

# Get the names for the model, once with and once without the covariances 
# included
parameter_names(
  my_model,
  dynamics = "anisotropic",
  covariance = "isotropic",
  parameters_only = FALSE
)

parameter_names(
  my_model,
  dynamics = "anisotropic",
  covariance = "isotropic",
  parameters_only = TRUE
)

}
