% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parameter_names_covariance}
\alias{parameter_names_covariance}
\title{Get the names for the parameters in the covariance matrix}
\usage{
parameter_names_covariance(d, covariance = "symmetric")
}
\arguments{
\item{d}{Integer denoting the dimensionality of the model.}

\item{covariance}{Character denoting the structure of the covariance matrix.
Can either be \code{"symmetric"} (symmetric around the diagonal) or 
\code{"isotropic"} (diagonal). Defaults to \code{"symmetric"}.}
}
\value{
Character vector containing the names for all parameters in the 
covariance matrix. The order corresponds to the assumptions of 
\code{\link[discounting]{get_bounds}} and \code{\link[discounting]{fill}}.
}
\description{
Get the names for the parameters in the covariance matrix
}
\examples{
parameter_names_covariance(2, covariance = "symmetric")

parameter_names_covariance(2, covariance = "isotropic")

}
