% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parameters}
\alias{parameters}
\alias{parameters,model-method}
\title{Extract model parameters}
\usage{
parameters(model, ...)

\S4method{parameters}{model}(model, vector = TRUE)
}
\arguments{
\item{model}{Instance of the \code{\link[discounting]{model-class}}}

\item{vector}{Logical denoting whether to output the parameters in a vector
(\code{TRUE}) or a list (\code{FALSE}). Defaults to \code{TRUE}.}
}
\value{
Numeric vector containing all parameters of the model.
}
\description{
Extract the parameters of a particular model. This includes the covariances,
so not only the deterministic model parameters.
}
\examples{
parameters(
  double_exponential(d = 2, k = 3),
  vector = TRUE
)

parameters(
  double_exponential(d = 2, k = 3),
  vector = FALSE
)

}
