% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict,exponential-method}
\alias{predict,exponential-method}
\alias{predict,quasi_hyperbolic-method}
\alias{predict,double_exponential-method}
\title{Compute the model expectations}
\usage{
\S4method{predict}{exponential}(object, data)

\S4method{predict}{quasi_hyperbolic}(object, data)

\S4method{predict}{double_exponential}(object, data)
}
\arguments{
\item{object}{Instance of the \code{\link[discounting]{model-class}}}

\item{data}{Instance of the \code{\link[discounting]{dataset-class}}}
}
\value{
Instance of the \code{\link[discounting]{dataset-class}} where the
slot \code{y} contains the predictions of the model
}
\description{
Within this package, the \code{predict} method computes the expectations of 
a particular model provided to \code{object} using the data provided to 
\code{data}. It plays a pivotal role in both the estimation and simulation of 
the models provided in this package.
}
\examples{
# Create an exponential discounting model
my_model <- exponential(
  parameters = list(
    "alpha" = 0,
    "beta" = 1,
    "gamma" = 0.5
  ),
  covariance = 1
)

# Create an instance of the dataset with only predictor values. It will throw
# a warning because Y is implied to be empty, but we capture this warning 
# with suppressWarnings()
data <- dataset(
  X = matrix(
    rnorm(10),
    nrow = 10,
    ncol = 1
  )
) |>
  suppressWarnings()

# Compute the values of Y as expected by the exponential discounting model 
# defined in my_model
predict(
  my_model,
  data
)

}
