% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{quasi_hyperbolic}
\alias{quasi_hyperbolic}
\title{Constructor for the \code{\link[discounting]{quasi_hyperbolic-class}}}
\usage{
quasi_hyperbolic(
  d = NA,
  k = NA,
  parameters = list(alpha = 0, beta = matrix(0, nrow = 1, ncol = 1), nu = matrix(0, nrow
    = 1, ncol = 1), kappa = matrix(0, nrow = 1, ncol = 1)),
  covariance = matrix(0, nrow = 1, ncol = 1),
  cholesky = FALSE
)
}
\arguments{
\item{d}{Integer denoting the number of dimensions of the model. Defaults to 
\code{NA}, in which case this dimensionality is inferred from the parameters.}

\item{k}{Integer denoting the number of independent variables. Defaults to 
\code{NA}, in which case this dimensionality is inferred from the parameters.}

\item{parameters}{List containing the parameters relevant to a particular 
model instance. Defaults to an empty model with \eqn{d = 1} and \eqn{k = 1}.}

\item{covariance}{Numeric matrix denoting the residual covariance of the 
model. Default to a matrix of \code{0}s with the dimensionality implied by 
\code{d}.}

\item{cholesky}{Logical denoting whether \code{covariance} is a 
lower-triangular decomposition matrix instead of an actual covariance matrix.
Defaults to \code{FALSE}.}
}
\value{
Instance of \code{\link[discounting]{quasi_hyperbolic-class}}
}
\description{
Defines an instance of the \code{\link[discounting]{quasi_hyperbolic-class}}, 
that is the class defining the quasi-hyperbolic discounting model. For the 
mathematical equations and how to specify them in this constructor, users can 
look at the details. For more information on the model itself, see the 
vignette on the model definitions.
}
\details{
The quasi-hyperbolic discounting model assumes that the effect of stimuli on 
affect fades away at a quasi-hyperbolic rate, that is first decreasing by a 
certain proportion after the first iteration and then decreasing at an 
exponential rate. The model can be defined as:

\deqn{\boldsymbol{y}_t = \boldsymbol{\alpha} + \sum_{j = 0}^t N^j K^i(j) B
\boldsymbol{x}_{t - j} + \boldsymbol{\epsilon}_t}

where \eqn{i} is an indicator function that is \eqn{0} when \eqn{j = 0} and 
\eqn{1} otherwise, \eqn{\boldsymbol{\alpha}} is a \eqn{d}-dimensional vector 
representing the mean of the process \eqn{\boldsymbol{y}}, \eqn{N} and 
\eqn{K} are \eqn{d \times d} matrices containing the forgetting factors, 
determining how long the effect of the independent variables 
\eqn{\boldsymbol{x}} on the process \eqn{\boldsymbol{y}} lingers on and 
therefore determining the dynamics of the system, \eqn{B} is a 
\eqn{d \times k}matrix containing the slopes of the independent variables 
\eqn{\boldsymbol{x}}, and \eqn{\boldsymbol{\epsilon}} represents the 
residuals of the system. Within this package, we assume that:

\deqn{\boldsymbol{\epsilon} \overset{iid}{\sim} N(\boldsymbol{0}, \Sigma)}

where \eqn{\Sigma} is a \eqn{d \times d} matrix representing the residual 
covariance matrix. 

The quasi-hyperbolic discounting model can also be written as an instance of 
the \eqn{VARMAX(1, 1)}, so that:

\deqn{\boldsymbol{y}_t = (I_d - N) \boldsymbol{\alpha} + B \boldsymbol{x}_t
N (K - I_d) B \boldsymbol{x}_{t - 1} + N \boldsymbol{y}_{t - 1} - 
N \boldsymbol{\epsilon}_{t - 1} + \boldsymbol{\epsilon}_t}

where \eqn{I_d} is a \eqn{d \times d} identity matrix.

Interestingly, in multiple dimensions (\eqn{d > 1}), there is an alternative 
definition of the quasi-hyperbolic discounting model, specifically one where 
the matrices \eqn{N} and \eqn{K} are flipped, so that:

\deqn{\boldsymbol{y}_t = \boldsymbol{\alpha} + \sum_{j = 0}^t \K^i(j) N^j B
\boldsymbol{x}_{t - j} + \boldsymbol{\epsilon}_t}

Its \eqn{VARMAX(1, 1)} representation then becomes:

<TO DO>

For simplicity, we only use the first definition of the quasi-hyperbolic model
within this package. 

To define the model, one should minimally define the parameters of the model
through \code{parameters} and \code{covariance}. In \code{covariance}, one 
either provides the \eqn{d \times d} residual covariance matrix \eqn{\Sigma}
(\code{cholesky = FALSE}) or a lower-triangular \eqn{d \times d} matrix 
\eqn{G} that represents part of the decomposition of this matrix 
(\code{cholesky = TRUE}), so that we can retrieve the covariance matrix as:

\deqn{\Sigma = G G^T}

The latter option ensures that the covariance matrix \eqn{\Sigma} defined 
within the model is positive-definite, but may require some additional 
thinking on the side of the user.

In \code{parameters}, one should provide a named list with instances 
\code{"alpha"}, \code{"beta"}, \code{"nu"}, and \code{"kappa"}, each defining 
the respective parameters \eqn{\boldsymbol{\alpha}}, \eqn{B}, \eqn{N} and 
\eqn{K} of the model. Note that the dimensionalities of these parameters 
should match up and are actively checked when constructing this class. 

There are some important restrictions to take into account when defining the 
model:
\itemize{
  \item{The eigenvalues \eqn{\lambda_i} of \eqn{N} and \eqn{K} should
        all lie between 0 and 1, or \eqn{\lambda_i \in [0, 1)}, a restriction
        that imposes exponential decay rather than sawtooth or more complex
        dynamics}
  \item{The covariance matrix \eqn{\Sigma} should be positive definite, as 
        explained above}
}
}
\examples{
quasi_hyperbolic(
  d = 2,
  k = 2,
  parameters = list(
    "alpha" = numeric(2),
    "beta" = diag(2) * 2,
    "nu" = diag(2) * 0.5,
    "kappa" = diag(2) * 0.75
  ),
  covariance = diag(2)
)

}
\seealso{
\code{\link[discounting]{model-class}}
\code{\link[discounting]{exponential-class}}
\code{\link[discounting]{quasi_hyperbolic-class}}
}
