% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recovery.R
\name{recovery}
\alias{recovery}
\alias{recovery,model-method}
\title{Simulate and estimate a model}
\usage{
recovery(sim_model, ...)

\S4method{recovery}{model}(
  sim_model,
  fit_model = NULL,
  iterations = 100,
  fx = list(),
  X = NULL,
  Xfun = NULL,
  N = NULL,
  dynamics = "isotropic",
  covariance = "symmetric",
  sim_dynamics = dynamics,
  sim_covariance = covariance,
  fit_dynamics = dynamics,
  fit_covariance = covariance,
  ...
)
}
\arguments{
\item{sim_model}{An instance of the \code{\link[discounting]{model-class}}
with which the data will be generated/simulated. Note that this model can 
be empty: Parameters will be simulated automatically within this function.}

\item{...}{Arguments passed on to \code{\link[discounting]{fit}}.}

\item{fit_model}{An instance of the \code{\link[discounting]{model-class}}
that will be estimated on the simulated data. Note that this model can be 
empty, but should have a same dimensionality (that is, \code{d} and \code{k})
as \code{sim_model}. By default, \code{fit_model} is equal to \code{sim_model}}

\item{iterations}{Integer denoting the number of iterations to run for the 
recovery/distinguishability study. Defaults to \code{100}.}

\item{fx}{Named list of functions that take the output of the estimation 
procedure and summarize it in a way. The function should only take in a 
single argument, it being the output of\code{\link[discounting]{fit}}, and 
should return a single value. Examples of functions that may be useful to 
include are estimates of \eqn{AIC} and \eqn{BIC}. Defaults to an empty list.}

\item{X}{Numeric vector or numeric matrix containing the values of the 
predictor variable(s). If a numeric vector, the dimension \code{k} of the 
predictor variables within \code{object} should be equal to \code{1}. If 
a numeric matrix, the dimension \code{k} of the predictor variables within 
\code{object} should be equal to the number of columns provided to this 
argument. The length of the provided vector or the number of rows of the 
provided matrix will determine the number of simulated datapoints. Defaults
to \code{NULL}. If provided, it overrides the value of \code{Xfun}.}

\item{Xfun}{Function or a list of functions that define how the values of 
the predictor variable(s) should be generated. The functions should take in 
a single argument \code{N} defining the number of values generated for this 
predictor variable, and hence the number of datapoints that should be 
generated for the simulation. If a list, then the result of each function in
the list will be bound together in a matrix. Defaults to \code{NULL}.}

\item{N}{Number of datapoints to simulate. Passed on to \code{Xfun} when 
defined. Defaults to \code{NULL}.}

\item{dynamics, sim_dynamics, fit_dynamics}{Character denoting the structure of 
the dynamic parameters of the models. See \code{\link[discounting]{fill}} for 
guidance on their potential values. By default, both the simulation and 
fitting model are \code{"isotropic"}.}

\item{covariance, sim_covariance, fit_covariance}{Character denoting the 
structure of the covariance matrix for the models. See 
\code{\link[discounting]{fill}} for guidance on their potential values. By
default, both the simulation and fitting model are \code{"symmetric"}.}
}
\value{
Named list containing a \code{data.frame} with simulated parameters
(under \code{"simulate"}) and a \code{data.frame} with estimated parameters, 
the result of the objective function, and other statistics that users can 
compute based on the output of \code{\link[discounting]{fit}}.
}
\description{
This function takes in one or two models of a same dimensionality and 
simulates data with the first and estimates the second model on those data.
This allows users to perform recovery studies -- the topic of which is to 
find out how accurately the parameters a model can be estimated -- and to 
perform distinguishability studies -- the topic of which is to find out how
well we can distinguish between one or the other model based on their fit
or some statistics.
}
\examples{
# Create a study in which we simulate from a slightly more complicated model
# and estimate a simpler model, both based on the exponential discounting 
# model
recovery(
  exponential(d = 1, k = 1),

  # Specifications of the recovery
  iterations = 10,
  fx = list(
    "aic" = function(x) 
      (length(x$residuals) * log(x$objective) / length(x$residuals) + 2 * length(x$parameters))
  ),

  # Arguments for simulate
  Xfun = \(x) rnorm(x),
  N = 25, 
  sim_dynamics = "symmetric",
  sim_covariance = "symmetric",

  # Arguments for fit
  fit_dynamics = "isotropic",
  fit_covariance = "isotropic",
  itermax = 25,
  trace = FALSE
)

}
