% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate,model-method}
\alias{simulate,model-method}
\title{Method for simulating from a model}
\usage{
\S4method{simulate}{model}(object, X = NULL, Xfun = NULL, N = NULL)
}
\arguments{
\item{object}{An instance of the \code{\link[discounting]{model-class}} 
defining the model}

\item{X}{Numeric vector or numeric matrix containing the values of the 
predictor variable(s). If a numeric vector, the dimension \code{k} of the 
predictor variables within \code{object} should be equal to \code{1}. If 
a numeric matrix, the dimension \code{k} of the predictor variables within 
\code{object} should be equal to the number of columns provided to this 
argument. The length of the provided vector or the number of rows of the 
provided matrix will determine the number of simulated datapoints. Defaults
to \code{NULL}. If provided, it overrides the value of \code{Xfun}.}

\item{Xfun}{Function or a list of functions that define how the values of 
the predictor variable(s) should be generated. The functions should take in 
a single argument \code{N} defining the number of values generated for this 
predictor variable, and hence the number of datapoints that should be 
generated for the simulation. If a list, then the result of each function in
the list will be bound together in a matrix. Defaults to \code{NULL}.}

\item{N}{Number of datapoints to simulate. Passed on to \code{Xfun} when 
defined. Defaults to \code{NULL}.}
}
\value{
An instance of the \code{\link[discounting]{dataset-class}} containing
the simulated data.
}
\description{
Takes in an instance of the \code{\link[discounting]{model-class}} and 
simulates data according to this model and the specifications of the user.
Depends heavily on the trajectories created by 
\code{\link[discounting]{predict}}.
}
\examples{
# Define the model
my_model <- exponential(
  parameters = list(
    "alpha" = numeric(2), 
    "beta" = matrix(1:2, ncol = 1),
    "gamma" = diag(2) * 0.75
  ),
  covariance = diag(2)
)

# Simulate data for this model using a function for generating X
simulate(
  my_model,
  Xfun = \(x) rnorm(x, mean = 0, sd = 1),
  N = 10
)

# Simulate data for this model using predefined values of X
simulate(
  my_model,
  X = rnorm(10, mean = 0, sd = 1)
)

}
