% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill}
\alias{fill}
\alias{fill,exponential-method}
\alias{fill,quasi_hyperbolic-method}
\alias{fill,double_exponential-method}
\title{Change model parameters}
\usage{
fill(model, parameters, ...)

\S4method{fill}{exponential}(
  model,
  parameters,
  dynamics = "isotropic",
  covariance = "symmetric",
  parameters_only = TRUE,
  cholesky = TRUE
)

\S4method{fill}{quasi_hyperbolic}(
  model,
  parameters,
  dynamics = "isotropic",
  covariance = "symmetric",
  parameters_only = TRUE,
  cholesky = TRUE
)

\S4method{fill}{double_exponential}(
  model,
  parameters,
  dynamics = "isotropic",
  covariance = "symmetric",
  parameters_only = TRUE,
  cholesky = TRUE
)
}
\arguments{
\item{model}{Instance of the \code{\link[discounting]{model-class}}}

\item{parameters}{Numeric vector containing the values of the parameters 
that should be assigned to the model. Importantly, the number of parameters
within this vector should correspond exactly to the number of parameters 
needed by the model. Ideally, you let functions internal to this package
handle this for you, rather than you defining these parameters manually.}

\item{...}{Additional arguments passed on to the methods.}

\item{dynamics}{Character denoting the structure of the dynamical matrices.
Can either be \code{"anisotropic"} (completely free), \code{"symmetric"}
(symmetric around the diagonal), and \code{"isotropic"} (diagonal). Note that
this influences different parameters for different models, namely 
\eqn{\Gamma} for the exponential discounting model, \eqn{N} and \eqn{K} for
the quasi-hyperbolic discounting model, and \eqn{\Gamma} and \eqn{N} for the
double-exponential discounting model. Defaults to \code{"isotropic"}.}

\item{covariance}{Character denoting the structure of the covariance matrix.
Can either be \code{"symmetric"} (symmetric around the diagonal) or 
\code{"isotropic"} (diagonal). Defaults to \code{"symmetric"}.}

\item{parameters_only}{Logical denoting whether to only fill the 
parameters in de \code{parameter} slot of the model (\code{TRUE}), or to 
fill the covariance matrix as well (\code{FALSE}). Defaults to \code{TRUE}.}

\item{cholesky}{Logical denoting whether the values of the covariance matrix 
should be taken as the values of its Cholesky decomposition instead. Defaults 
to \code{TRUE}.}
}
\value{
Instance of the \code{\link[discounting]{model-class}} containing 
the values \code{parameters} in its \code{parameter}-slot.
}
\description{
Fill the model parameters of a particular instance of the 
\code{\link[discounting]{model-class}} with the ones provided in a numeric
vector. Used in \code{\link[discounting]{objective_function}} to allow for
predictions with the model while cohering to the input typical numerical 
optimizers such as \code{\link[DEoptim]{DEoptim}} and 
\code{\link[nloptr]{nloptr}} provide within the objective function.
}
\examples{
# Create an empty instance of the exponential discounting model
my_model <- exponential()

# Assign the parameters of a vector to this exponential function
fill(
  my_model,
  c(1, 2, 0.5)
)

}
