% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill_covariance}
\alias{fill_covariance}
\title{Fill the covariance matrix with values}
\usage{
fill_covariance(d, parameters, covariance = "symmetric", cholesky = TRUE)
}
\arguments{
\item{d}{Integer denoting the dimensionality of the model.}

\item{parameters}{Numeric vector containing the values of the parameters 
that should be assigned to the model. Importantly, the number of parameters
within this vector should correspond exactly to the number of parameters 
needed by the model. Ideally, you let functions internal to this package
handle this for you, rather than you defining these parameters manually.}

\item{covariance}{Character denoting the structure of the covariance matrix.
Can either be \code{"symmetric"} (symmetric around the diagonal) or 
\code{"isotropic"} (diagonal). Defaults to \code{"symmetric"}.}

\item{cholesky}{Logical denoting whether the values of the covariance matrix 
should be taken as the values of its Cholesky decomposition instead. Defaults 
to \code{TRUE}.}
}
\value{
Matrix filled with the specified values.
}
\description{
Fill the covariance matrix with values
}
\examples{
fill_covariance(
  2,
  1:3,
  covariance = "symmetric",
  cholesky = FALSE
)

fill_covariance(
  2,
  1:2,
  covariance = "isotropic",
  cholesky = FALSE
)

}
