% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\alias{fit,model,dataset-method}
\title{Fit a model to data}
\usage{
fit(model, data, ...)

\S4method{fit}{model,dataset}(
  model,
  data,
  dynamics = "isotropic",
  covariance = "symmetric",
  optimizer = "DEoptim",
  ...
)
}
\arguments{
\item{model}{Instance of the \code{\link[discounting]{model-class}}, 
defining the model to evaluate the objective function for.}

\item{data}{Instance of the \code{\link[discounting]{dataset-class}}
containing the data to fit the model to.}

\item{...}{Arguments passed on to the control parameters of the optimizer, 
either to \code{\link[DEoptim]{DEoptim.control}} or the \code{opts} 
argument of \code{\link[nloptr]{nloptr}}.}

\item{dynamics}{Character denoting the structure of the dynamical matrices.
Can either be \code{"anisotropic"} (completely free), \code{"symmetric"}
(symmetric around the diagonal), and \code{"isotropic"} (diagonal). Note that
this influences different parameters for different models, namely 
\eqn{\Gamma} for the exponential discounting model, \eqn{N} and \eqn{K} for
the quasi-hyperbolic discounting model, and \eqn{\Gamma} and \eqn{N} for the
double-exponential discounting model. Defaults to \code{"isotropic"}.}

\item{covariance}{Character denoting the structure of covariance matrix.
Can either by \code{"symmetric"} (symmetric around the diagonal) and 
\code{"isotropic"} (diagonal). Defaults to \code{"symmetric"}.}

\item{optimizer}{Character denoting the optimizer to use for the estimation.
Can either be \code{"DEoptim"} for the differential evolution algorithm in 
\code{\link[DEoptim]{DEoptim}} or \code{"nloptr"} for the library implemented
in \code{\link[nloptr]{nloptr}}. Defaults to \code{"DEoptim"}.}
}
\value{
An named list containing an instance of the 
\code{\link[discounting]{model-class}} with the estimated parameters 
(\code{"model"}), the results of the optimization procedure (\code{"fit"}), 
the value of the objective after ending the optimization procedure 
(\code{"objective"}), the residuals of the model (\code{"residuals"}), and a 
named vector containing the values of the estimated parameters linked to a 
character vector explaining their content (\code{"parameters"}).
}
\description{
This method enables the user to fit a model (defined in \code{model}) on a
particular dataset (defined in \code{data}). The estimation procedure makes
use of numerical optimization using either \code{\link[DEoptim]{DEoptim}} or
\code{\link[nloptr]{nloptr}}, as specified by the user. Estimation proceeds
through an optimization according to the output of the objective function 
of the provided model, as defined through 
\code{\link[discounting]{objective_function}}, thus using least-squares
as optimization standard.
}
\details{
Note that currently, least-squares estimation is assumed, meaning that the 
covariance matrix is left out of the objective function. If maximum-likelihood
were needed instead, then this function would need to change.
}
\examples{
# Simulate data to use for this example
data <- simulate(
  quasi_hyperbolic(
    parameters = list(
      "alpha" = c(1, -1) ,
      "beta" = matrix(2, nrow = 2, ncol = 2),
      "nu" = diag(2) * 0.75,
      "kappa" = diag(2) * 0.5
    ),
    covariance = matrix(c(1, 0.25, 0.25, 1), nrow = 2, ncol = 2)
  ),
  X = matrix(rnorm(200), nrow = 100, ncol = 2)
)

# Evaluate the objective function for an exponential model with a particular
# set of parameters
fit(
  exponential(d = 2, k = 2),
  data,
  dynamics = "isotropic",
  covariance = "isotropic",
  itermax = 50,
  trace = FALSE
)

}
