% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parameters}
\alias{parameters}
\alias{parameters,model-method}
\title{Extract model parameters}
\usage{
parameters(model, ...)

\S4method{parameters}{model}(model, parameters_only = FALSE, ...)
}
\arguments{
\item{model}{Instance of the \code{\link[discounting]{model-class}}}

\item{...}{Additional arguments passed on to \code{\link[discounting]{index}}}

\item{parameters_only}{Logical denoting whether to only extract the 
determinstic parameters of the model (\code{TRUE}) or also include the 
covariance parameters (\code{FALSE}). Defaults to \code{FALSE}.}
}
\value{
Numeric vector containing all parameters of the model.
}
\description{
Extract the parameters of a particular model.
}
\examples{
parameters(double_exponential(d = 2, k = 3))

parameters(double_exponential(d = 2, k = 3))

}
